[fix] Updates for numpy 2.*

+++ world/grid_world.py
@@ -132,7 +132,7 @@ class GridWorld:
         return spawned
 
     def _build_map(self):
-        self.map = np.zeros((self.base_map.shape[0], self.base_map.shape[1], 2), np.int)
+        self.map = np.zeros((self.base_map.shape[0], self.base_map.shape[1], 2), np.int32)
         self.preys.clear()
         for sc in self.team_spawn_coordinates:
             sc.clear()
@@ -142,14 +142,14 @@ class GridWorld:
         for x in range(self.base_map.shape[1]):
             for y in range(self.base_map.shape[0]):
                 if self.base_map[y, x] == -1:
-                    self.map[y, x] = np.array([-1, -1], dtype=np.long) # Unpassable tiles
+                    self.map[y, x] = np.array([-1, -1], dtype=np.float64) # Unpassable tiles
                     continue # Passable tile
                 elif self.base_map[y, x] == -2:
-                    self.map[y, x] = np.array((self.playable_teams_num, len(self.preys)), dtype=np.long)
+                    self.map[y, x] = np.array((self.playable_teams_num, len(self.preys)), dtype=np.float64)
                     self.preys.append(Entity(x, y, len(self.preys), self.playable_teams_num, True))
                     continue
                 elif 0 <= self.base_map[y, x] <= self.playable_teams_num:
-                    self.map[y, x] = np.array([-1, 0], dtype=np.long)
+                    self.map[y, x] = np.array([-1, 0], dtype=np.float64)
                     if 0 < self.base_map[y, x] <= self.playable_teams_num:
                         self.team_spawn_coordinates[self.base_map[y, x]-1].append((x, y))  # Add Spawn point
                 else:
